initial gsettings for evolution mail (not automatically run)
the first setting (thread-expand true) is the default behavior.
but is placed there to remind self / others why I used that setting
instead of false.

subsequent settings can be added to this script.

note that this script is (currently) not run automatically.
This may change, depending if some gsettings for evolution need to be
tweaked away from their defaults